Replace review audit CLI base with utility
