--- scinoephile/cli/audit/audit_review_cli_base.py
@@ -1,42 +1,0 @@
-#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
-#  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Shared command-line support for subtitle review audits."""
-
-from __future__ import annotations
-
-from argparse import ArgumentParser
-from collections.abc import Sequence
-from pathlib import Path
-
-from scinoephile.core.llms import TestCase
-from scinoephile.llms.review import ReviewManager
-
-from .audit_cli_base import AuditCliBase
-
-__all__ = ["AuditReviewCliBase"]
-
-
-class AuditReviewCliBase(AuditCliBase):
-    """Shared command-line support for subtitle review audits."""
-
-    @staticmethod
-    def load_review_test_cases(
-        parser: ArgumentParser,
-        json_path: Path | None,
-    ) -> Sequence[TestCase]:
-        """Load optional review test cases from JSON.
-
-        Arguments:
-            parser: parser used to report user-facing errors
-            json_path: optional review JSON path
-        Returns:
-            loaded review test cases
-        """
-        if json_path is None:
-            return ()
-        return AuditCliBase.load_test_cases(
-            parser,
-            json_path,
-            ReviewManager,
-            workflow_name="review",
-        )
